requirements.txt
Flask==2.3.3
qrcode==7.4.2
Pillow==10.0.1
byteplus-python-sdk-v2
python-dotenv==1.0.0
